+++ src/main/java/frc/robot/BuildConstants.java
@@ -7,14 +7,13 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "Robot2024-1";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 48;
-  public static final String GIT_SHA = "6dab8608750c7defed574f8811f305961222d588";
-  public static final String GIT_DATE = "2023-10-09 19:45:23 EDT";
-  public static final String GIT_BRANCH = "EricSSubsystemTransfer_10_9";
-  public static final String BUILD_DATE = "2023-10-12 18:24:44 EDT";
-  public static final long BUILD_UNIX_TIME = 1697149484664L;
-  public static final int DIRTY = 0;
-
+  public static final int GIT_REVISION = 54;
+  public static final String GIT_SHA = "180c257b6015ae29f39527f3df35cd95cbef1cfc";
+  public static final String GIT_DATE = "2023-10-12 18:29:11 EDT";
+  public static final String GIT_BRANCH = "JTM_MovingConstants";
+  public static final String BUILD_DATE = "2023-10-12 19:46:20 EDT";
+  public static final long BUILD_UNIX_TIME = 1697154380792L;
+  public static final int DIRTY = 1;
 
   private BuildConstants(){}
 }
