+++ src/main/java/frc/robot/swerve/SwerveModule.java
@@ -12,7 +12,7 @@ import com.revrobotics.SparkMaxPIDController;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
-import frc.robot.Constants.SwerveModuleConstants;
+import frc.robot.RobotConstants.SwerveModuleConstants;
 import frc.robot.sensors.ThriftyEncoder;
 
 /**
